src/test/javascript/spec/app/entities/commodity/commodity.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_FORMAT, DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { CommodityService } from 'app/entities/commodity/commodity.service';
import { ICommodity, Commodity } from 'app/shared/model/commodity.model';

describe('Service Tests', () => {
  describe('Commodity Service', () => {
    let injector: TestBed;
    let service: CommodityService;
    let httpMock: HttpTestingController;
    let elemDefault: ICommodity;
    let expectedResult;
    let currentDate: moment.Moment;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = {};
      injector = getTestBed();
      service = injector.get(CommodityService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new Commodity(0, currentDate, 'AAAAAAA', 0, false, 0, 0, currentDate);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            dateOfLastStockCheck: currentDate.format(DATE_FORMAT),
            createdDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: elemDefault });
      });

      it('should create a Commodity', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            dateOfLastStockCheck: currentDate.format(DATE_FORMAT),
            createdDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );
        const expected = Object.assign(
          {
            dateOfLastStockCheck: currentDate,
            createdDate: currentDate
          },
          returnedFromService
        );
        service
          .create(new Commodity(null))
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should update a Commodity', () => {
        const returnedFromService = Object.assign(
          {
            dateOfLastStockCheck: currentDate.format(DATE_FORMAT),
            name: 'BBBBBB',
            amount: 1,
            perishable: true,
            amountExpirableInNext3months: 1,
            createdByUserId: 1,
            createdDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            dateOfLastStockCheck: currentDate,
            createdDate: currentDate
          },
          returnedFromService
        );
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should return a list of Commodity', () => {
        const returnedFromService = Object.assign(
          {
            dateOfLastStockCheck: currentDate.format(DATE_FORMAT),
            name: 'BBBBBB',
            amount: 1,
            perishable: true,
            amountExpirableInNext3months: 1,
            createdByUserId: 1,
            createdDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );
        const expected = Object.assign(
          {
            dateOfLastStockCheck: currentDate,
            createdDate: currentDate
          },
          returnedFromService
        );
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Commodity', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
